+++ cmd/hidden_lake/loader/internal/config/init.go
@@ -4,6 +4,8 @@ import (
 	"os"
 
 	hll_settings "github.com/number571/go-peer/cmd/hidden_lake/loader/pkg/settings"
+	hlt_settings "github.com/number571/go-peer/cmd/hidden_lake/traffic/pkg/settings"
+	logger "github.com/number571/go-peer/internal/logger/std"
 )
 
 func InitConfig(cfgPath string, initCfg *SConfig) (IConfig, error) {
@@ -12,6 +14,13 @@ func InitConfig(cfgPath string, initCfg *SConfig) (IConfig, error) {
 	}
 	if initCfg == nil {
 		initCfg = &SConfig{
+			FSettings: &SConfigSettings{
+				FMessagesCapacity: hlt_settings.CDefaultMessagesCapacity,
+			},
+			FLogging: []string{logger.CLogInfo, logger.CLogWarn, logger.CLogErro},
+			FAddress: &SAddress{
+				FHTTP: hll_settings.CDefaultHTTPAddress,
+			},
 			FProducers: []string{hll_settings.CDefaultProducerAddress},
 			FConsumers: []string{hll_settings.CDefaultConsumerAddress},
 		}