+++ cmd/hidden_lake/loader/internal/config/types.go
@@ -6,6 +6,7 @@ type IConfig interface {
 	GetLogging() logger.ILogging
 	GetSettings() IConfigSettings
 
+	GetAddress() IAddress
 	GetProducers() []string
 	GetConsumers() []string
 }
@@ -14,4 +15,9 @@ type IConfigSettings interface {
 	GetNetworkKey() string
 	GetWorkSizeBits() uint64
 	GetMessagesCapacity() uint64
+}
+
+type IAddress interface {
+	GetHTTP() string
+	GetPPROF() string
 }