+++ cmd/hidden_lake/loader/pkg/client/types.go
@@ -1,0 +1,15 @@
+package client
+
+type IClient interface {
+	GetIndex() (string, error)
+
+	RunTransfer() error
+	StopTransfer() error
+}
+
+type IRequester interface {
+	GetIndex() (string, error)
+
+	RunTransfer() error
+	StopTransfer() error
+}